Allow NameIDType objects to be seen in a JSON output
The admin/test page converts the object to JSON and objects with
protected attributes to not get shown. This allows those objects to
appear on the debug page.

This will also allow the contents of NameIDTypes to be seen in any
JSON serialization. So if there are thing that we want to mask for any
reason then we should look deeper into this. Though if that is the
case then the code calling json_encode should itself probably be
copying and scrubbing the data to serialize if there are things we do
not want to have in the output.